requirements-client.txt
requests==2.31.0
websocket-client==1.6.1
aiohttp==3.8.5
websockets==11.0.3
prometheus-client==0.17.1 